Fix Firefox menu bug causing an unwanted line break.
Firefox has a known bug where floated boxes can cause line breaks.
In this case, the right floated menu items are creating a line break
in the menu and obscuring some of the content. Create a simple fix by
listing the floated element first in the JSX so thatthe browser
positions it before rendering the elements still in the page flow.